requirements.txt
prophet
scikit-learn
python-binance
uvicorn
uvloop
httptools
pymongo
pandas
dash
pydantic
fastapi
httpx